project/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@292954b547b838612c2921feba59aa999116bb83
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
bitsandbytes==0.44.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
debugpy==1.8.7
decorator==5.1.1
dill==0.3.8
duckduckgo_search==6.3.0
einops==0.8.0
executing==2.1.0
faiss-cpu==1.9.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.134
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.9.3
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
ollama==0.3.3
openai-clip==1.0.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft @ git+https://github.com/huggingface/peft.git@c039b00358c1c7d9139f92c510fb035ebfe753fa
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
primp==0.6.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
qrdet==2.5
qreader==3.14
quadrilateral-fitter==1.12
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.14.1
sentence-transformers==3.2.0
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@fa3f2db5c7405a742fcb8f686d3754f70db00977
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.11
ultralytics-thop==2.0.9
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.15.2
